# specs/001-auth-fullstack-core/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API - Phase I
  description: |
    Multi-user Todo API with JWT authentication.
    All endpoints require valid JWT token in Authorization header.
    Users can only access and modify their own tasks.
  version: 1.0.0
  contact:
    name: Todo Phase 2 Project

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todo.example.com
    description: Production server

tags:
  - name: Tasks
    description: Task CRUD operations (all require authentication)

security:
  - bearerAuth: []

paths:
  /api/{user_id}/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks for a user
      description: |
        Returns all tasks owned by the specified user.
        Requires JWT token with matching user_id.
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: |
        Creates a new task for the specified user.
        Requires JWT token with matching user_id.
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/{user_id}/tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get a specific task
      description: |
        Returns a single task by ID.
        Requires JWT token with matching user_id.
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Tasks
      summary: Update a task
      description: |
        Updates the title of an existing task.
        Requires JWT token with matching user_id.
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: |
        Permanently deletes a task.
        Requires JWT token with matching user_id.
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/taskId'
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      tags:
        - Tasks
      summary: Toggle task completion status
      description: |
        Toggles the is_completed status of a task.
        If currently incomplete, marks as complete.
        If currently complete, marks as incomplete.
        Requires JWT token with matching user_id.
      operationId: toggleTaskComplete
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth.
        Include in Authorization header as: Bearer <token>

  parameters:
    userId:
      name: user_id
      in: path
      required: true
      description: UUID of the user (must match JWT token subject)
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"

    taskId:
      name: task_id
      in: path
      required: true
      description: UUID of the task
      schema:
        type: string
        format: uuid
        example: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        user_id:
          type: string
          format: uuid
          description: Owner user identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          maxLength: 500
          description: Task description
          example: "Buy groceries"
        is_completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2026-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-01-15T10:30:00Z"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task description
          example: "Buy groceries"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: New task description
          example: "Buy groceries and milk"

    TaskListResponse:
      type: object
      required:
        - tasks
        - count
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        count:
          type: integer
          description: Total number of tasks
          example: 5

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Not authenticated"

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              summary: No token provided
              value:
                detail: "Not authenticated"
            invalid_token:
              summary: Token invalid or expired
              value:
                detail: "Invalid token"

    Forbidden:
      description: Authenticated but not authorized to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Forbidden"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not found"

    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Bad request"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            detail:
              - loc: ["body", "title"]
                msg: "field required"
                type: "value_error.missing"